feat: move conditimal interface
